+++ nichtparasoup/imagecrawler/__init__.py
@@ -5,6 +5,7 @@ from typing import Dict, Optional, Type
 from nichtparasoup.core.imagecrawler import BaseImageCrawler
 
 from .dummy import Dummy
+from .instagram import InstagramHashtag, InstagramProfile
 from .picsum import Picsum
 from .reddit import Reddit
 
@@ -12,6 +13,8 @@ _imagecrawlers = dict(
     Dummy=Dummy,
     Picsum=Picsum,
     Reddit=Reddit,
+    InstagramProfile=InstagramProfile,
+    InstagramHashtag=InstagramHashtag,
 )